Feat: Refactor kendo grid in grid-core component
