+++ projects/safe/src/lib/components/ui/grid-core/array-filter-menu/array-filter-menu.component.ts
