+++ projects/safe/src/lib/components/ui/grid-core/expanded-comment/expanded-comment.component.html
@@ -1,0 +1,11 @@
+<button mat-icon-button mat-dialog-close [mat-dialog-close]="data.comment" color="warn" class="close-settings">
+    <mat-icon>close</mat-icon>
+</button>
+<div mat-dialog-content class="mat-dialog-content">
+    <h1>{{data.title}}</h1>
+    <textarea id="comment" [(ngModel)]="comment" rows="15" class="modal-textinput" [readonly]="data.readOnly"></textarea>
+</div>
+<div mat-dialog-actions align="end">
+    <button mat-button [mat-dialog-close]="false">Close</button>
+    <button mat-flat-button [mat-dialog-close]="comment" cdkFocusInitial color="primary" [disabled]="data.readOnly">Save</button>
+</div>
