+++ projects/safe/src/lib/components/ui/grid-core/expanded-comment/expanded-comment.component.scss
