+++ projects/safe/src/lib/components/ui/grid-core/expanded-comment/expanded-comment.component.ts
@@ -1,9 +1,10 @@
-import { Component, EventEmitter, Inject, OnInit, Output } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
 
 interface DialogData {
   title: string;
   comment: string;
+  readOnly?: boolean;
 }
 
 @Component({
@@ -12,9 +13,6 @@ interface DialogData {
   styleUrls: ['./expanded-comment.component.scss']
 })
 export class SafeExpandedCommentComponent implements OnInit {
-
-  // === EMIT STEP CHANGE FOR WORKFLOW ===
-  @Output() goToNextStep: EventEmitter<any> = new EventEmitter();
 
   public comment = '';
 
@@ -27,7 +25,4 @@ export class SafeExpandedCommentComponent implements OnInit {
     this.comment = this.data.comment;
   }
 
-  // onClose(): void {
-  //   this.dialogRef.close({data: this.comment});
-  // }
 }
