+++ projects/safe/src/lib/components/ui/grid-core/models/grid-settings.model.ts
@@ -1,0 +1,25 @@
+import { CompositeFilterDescriptor } from '@progress/kendo-data-query';
+
+// TO-DO Finish it
+export interface GridSettings {
+    id?: string;
+    query?: {
+        name: string;
+        fields: any[];
+        sort?: {
+            field?: string;
+            order?: 'asc' | 'desc';
+        },
+        filter?: CompositeFilterDescriptor
+    };
+    actions?: {
+        delete?: boolean;
+        history?: boolean;
+        convert?: boolean;
+        update?: boolean;
+        inlineEdition?: boolean;
+    };
+    showDetails?: boolean;
+    showExport?: boolean;
+    defaultLayout?: any;
+}
